prepared data for calculating similarities
