+++ data/places.csv
@@ -91,7 +91,7 @@
 90,Centre de culture Ibn Mahrez,Centre de culture
 91,Centre de culture  Ibntachefine (michelet),Centre de culture
 92,Bibliotheque Cathedrale,Bibliotheque
-93,Bibiliotheque es seddikia,Bibiliotheque
+93,Bibliotheque es seddikia,Bibliotheque
 94,Bibliotheque Le petit lecteur ,Bibliotheque
 95,Theatre de Verdure,Theatre
 96,Theatre Regional,Theatre
